refactor: Move all tests to separate test_* files
- Move unit tests from main.rs to test_unit.rs
- Move async_concurrency_tests.rs to test_async_concurrency.rs
- Move property_tests.rs to test_property.rs
- Move LogLevel enum from main.rs to lib.rs and make it public
- Remove test module declarations from lib.rs and main.rs
- Update imports in moved test files
- All 33 tests continue to pass with new structure

Test files now follow consistent test_* naming pattern:
- test_cli.rs (4 tests)
- test_integration.rs (19 tests)
- test_unit.rs (3 tests)
- test_async_concurrency.rs (4 tests)
- test_property.rs (3 tests)
